EMBERS OF EDEN: Added 30+ custom effects, added Remembrance stave, added Annihilator axe, added Cataclysm blade, added Mourning Star, added Modular Curios, added Architect's Crucible improvement, added Eerie fuller, added Euclidean Dagger, added global loot modifiers, added heavy crossbow stave, added light crossbow stave, added light crossbow stock, added crossbow hones, added Reaping scythe, reworked Rending Scissors, added Shockwave hammer, added Slime hammer, added Soul Fire Scythe, added Thousand Cold Nights katana improvement, added Tonal blade, and added a bunch of items and new materials I'm not writing anything more this is too long

+++ src/main/java/net/acetheeldritchking/art_of_forging/effects/VengeanceEffect.java
@@ -46,7 +46,7 @@ public class VengeanceEffect {
             if (heldStack.getItem() instanceof ModularItem item)
             {
                 // Extra damage
-                int level = item.getEffectLevel(heldStack, vengeanceEffect);
+                double level = item.getEffectLevel(heldStack, vengeanceEffect);
 
                 // Hearts left
                 float eff = item.getEffectEfficiency(heldStack, vengeanceEffect);
